Removes unecessary add to cart
