--- src/components/shopping/AddToCart/AddToCartNotification.js
@@ -1,18 +1,0 @@
-// import React from 'react';
-// import { connect } from 'react-redux';
-// import { productSelectors } from '../../../ducks/products';
-//
-// class AddToCartNotification extends React.Component {
-//   render() {
-//     return (
-//       <div></div>
-//     );
-//   }
-// }
-//
-//
-// const mapState = ({ products }, { productId }) => ({
-//   productName: productSelectors.productNameFromId(products, productId),
-// });
-//
-// export default connect(mapState)(AddToCartNotification);
