--- src/components/shopping/AddToCart/AddToCartOptions/EmptyButton.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export default () => <div className="add-to-cart empty" />;
