--- src/components/shopping/AddToCart/AddToCartOptions/LoginButton.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Link } from '../../../elements';
-
-export default () => (
-  <Link className="add-to-cart login" to="LOGIN">
-    Login
-  </Link>
-);
