--- src/components/shopping/AddToCart/AddToCartOptions/UnavailableButton.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <div className="add-to-cart out-of-stock">Out of Stock</div>
-);
