+++ src/ducks/categories/actions.js
@@ -7,8 +7,8 @@ const getParentCategories = () => async (dispatch, getState, api) => {
 };
 
 const getTimeslot = () => async (dispatch, getState, api) => {
-  const res = await api.get('/cart/timeslots/default');
-  dispatch({ type: types.GET_TIMESLOT, payload: res.data });
+  // const res = await api.get('/cart/timeslots/default');
+  // dispatch({ type: types.GET_TIMESLOT, payload: res.data });
 };
 
 export default {
