+++ src/ducks/shopping/actions.js
@@ -1,8 +1,8 @@
 import types from './types';
 
 const getTimeslot = () => async (dispatch, getState, api) => {
-  const res = await api.get('/cart/timeslots/default');
-  dispatch({ type: types.GET_TIMESLOT, payload: res.data });
+  // const res = await api.get('/cart/timeslots/default');
+  // dispatch({ type: types.GET_TIMESLOT, payload: res.data });
 };
 
 const getTimeslots = () => async (dispatch, getState, api) => {
